requirements.txt
# =============================================================================
# Dependencias - API Iris v2 (Projeto Final)
# =============================================================================

# API Framework
fastapi==0.115.13
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20

# Autenticacao JWT
PyJWT==2.10.1

# Machine Learning
numpy>=2.0.0
scikit-learn>=1.6.0

# Rate Limiting (NOVO!)
slowapi==0.1.9

# Logs Estruturados (NOVO!)
python-json-logger==2.0.7

# Metricas Prometheus (NOVO!)
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.0

# HTTP Client (para testes)
httpx==0.28.1
